working on windows warnings
git-svn-id: https://svn.code.sf.net/p/teem/code/teem/trunk@3802 3d70eeeb-363e-0410-a505-8a46323a89f2

+++ src/limn/io.c
@@ -141,7 +141,7 @@ limnPolyDataWriteIV(FILE *file, const limnPolyData *pld) {
   unsigned int primIdx, xyzwIdx, rgbaIdx, normIdx, bitFlag,
     baseVertIdx;
   int haveStrips, haveTris, haveElse;
-  float xyz[3], norm[3], len;
+  double xyz[3], norm[3], len;
 
   if (!(file && pld)) {
     sprintf(err, "%s: got NULL pointer", me);
